code refactoring for Bsim DMA integration

+++ cpp/portal.h
@@ -98,9 +98,8 @@ class PortalMemory : public PortalRequest {
   int alloc(size_t size, PortalAlloc *portalAlloc);
   int reference(PortalAlloc* pa);
   virtual void sglist(unsigned long off, unsigned long long addr, unsigned long len) = 0;
-  virtual void paref(unsigned long off, unsigned long long ref) = 0;
+  virtual void paref(unsigned long off, unsigned long ref) = 0;
 };
-
 
 // ugly hack (mdk)
 typedef int SGListId;
